Converted Terminal to TypeScript
